eprescriptions-backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==2.0.0
click==8.3.2
cryptography==46.0.7
email-validator==2.2.0
fastapi==0.135.3
greenlet==3.4.0
h11==0.16.0
httptools==0.7.1
idna==3.11
PyJWT==2.10.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
